wikipedia-bias-analyzer/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.23.2
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.10.6
pydantic-settings==2.8.1

# Database
psycopg[binary]==3.2.5

# Environment and configuration
python-dotenv==1.0.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Task queue
chancy==0.18.0

# LLM integration
openai==1.2.4

# Testing
pytest==7.4.3
httpx==0.25.1
pytest-asyncio==0.23.5
pytest-mock==3.12.0

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1

# Wikipedia API
wikipedia-api==0.8.1

# Markdownify
markdownify==0.11.0

# SQLite for testing (usually included in Python)

